% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_project.R
\name{create_project}
\alias{create_project}
\title{create project}
\usage{
create_project(name = NULL, type = NULL)
}
\arguments{
\item{name}{name of the project}

\item{type}{create an app or a shiny-package}
}
\description{
create project
}
